Add the idea of a WebHDFS helper to handle some of the methods used during the plugin routines, later on some of this plugins should be refactored back in to allow better unit testing, but for now this makes the main plugin class a bit cleaner
